privacy: add auto and manual setup screens
