+++ app/src/main/AndroidManifest.xml
@@ -185,6 +185,14 @@
             android:name=".activities.SetupPrivacy"
             android:launchMode="singleTask" />
 
+        <activity
+            android:name=".activities.SetupMixerAccounts"
+            android:launchMode="singleTask" />
+
+        <activity
+            android:name=".activities.ManualMixerSetup"
+            android:launchMode="singleTask" />
+
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="${applicationId}.fileprovider"
